Move pipeline repo functions under repository
