+++ app/repository/pipeline.sql
@@ -112,34 +112,32 @@ group by family_import_id
 )
 
 SELECT
-f.title as "family_title",
-p.title as "physical_document_title",
-f.description as "family_description",
-CASE
-  WHEN f.family_category in ('UNFCCC', 'MCF') THEN UPPER(f.family_category::TEXT)
-  ELSE INITCAP(f.family_category::TEXT)
-END "family_category",
-fp.published_date as "family_published_date",
-d.import_id as "family_document_import_id",
-ds.name as "family_document_slug",
-f.import_id as "family_import_id",
-fs.name as "family_slug",
-p.source_url as "physical_document_source_url",
-INITCAP(d.valid_metadata::json#>>'{
-  type,0}') as
-"family_document_type",
-o.name as "organisation_name",
-geos.geographies as "geographies",
-c.import_id as "corpus_import_id",
-c.corpus_type_name as "corpus_type_name",
-langs.languages AS "languages",
-fm.value as "family_metadata",
-d.valid_metadata as "family_document_metadata"
+  f.title as "family_title",
+  p.title as "physical_document_title",
+  f.description as "family_description",
+  CASE
+    WHEN f.family_category in ('UNFCCC', 'MCF') THEN UPPER(f.family_category::TEXT)
+    ELSE INITCAP(f.family_category::TEXT)
+  END "family_category",
+  fp.published_date as "family_published_date",
+  d.import_id as "family_document_import_id",
+  ds.name as "family_document_slug",
+  f.import_id as "family_import_id",
+  fs.name as "family_slug",
+  p.source_url as "physical_document_source_url",
+  INITCAP(d.valid_metadata::json#>>'{
+    type,0}') as
+  "family_document_type",
+  o.name as "organisation_name",
+  geos.geographies as "geographies",
+  c.import_id as "corpus_import_id",
+  c.corpus_type_name as "corpus_type_name",
+  langs.languages AS "languages",
+  fm.value as "family_metadata",
+  d.valid_metadata as "family_document_metadata"
 FROM physical_document p
-JOIN family_document d
-ON p.id = d.physical_document_id
-JOIN family f
-ON d.family_import_id = f.import_id
+JOIN family_document d ON p.id = d.physical_document_id
+JOIN family f ON d.family_import_id = f.import_id
 FULL JOIN (
   SELECT
   family_geography.family_import_id as "family_import_id",
@@ -151,14 +149,10 @@ FULL JOIN (
   ON geography.id = family_geography.geography_id
   GROUP BY family_geography.family_import_id
 ) fg ON fg.family_import_id=f.import_id
-join family_corpus fc
-on f.import_id = fc.family_import_id
-join corpus c
-on fc.corpus_import_id = c.import_id
-join organisation o
-on c.organisation_id = o.id
-join family_metadata fm
-on fm.family_import_id = f.import_id
+join family_corpus fc on f.import_id = fc.family_import_id
+join corpus c on fc.corpus_import_id = c.import_id
+join organisation o on c.organisation_id = o.id
+join family_metadata fm on fm.family_import_id = f.import_id
 LEFT OUTER JOIN (
     SELECT family_document.import_id AS family_document_import_id,
            json_agg(distinct(language.name)) AS languages
@@ -175,9 +169,10 @@ LEFT OUTER JOIN (
 	JOIN geography ON geography.id = family_geography.geography_id
 	GROUP BY family_geography.family_import_id
 ) AS geos ON geos.family_import_id = f.import_id
-LEFT JOIN most_recent_doc_slugs ds
-on ds.family_document_import_id = d.import_id
+LEFT JOIN most_recent_doc_slugs ds on ds.family_document_import_id = d.import_id
 LEFT JOIN most_recent_family_slugs fs on fs.family_import_id = f.import_id
 LEFT JOIN event_dates fp on fp.family_import_id = f.import_id
-WHERE d.document_status != 'DELETED' AND fg.family_import_id = f.import_id
-ORDER BY d.last_modified desc, d.created desc, d.ctid desc, f.import_id
+WHERE 
+  d.document_status != 'DELETED' AND fg.family_import_id = f.import_id
+ORDER BY
+  d.last_modified desc, d.created desc, d.ctid desc, f.import_id
